fix: correccion al elegir modulacion xd

+++ src/main/java/py/una/pol/rest/model/ModulationInfo.java
@@ -2,7 +2,7 @@ package py.una.pol.rest.model;
 
 public class ModulationInfo {
     public String modulation;
-    public int distance;
+    public double distance;
 
     public ModulationInfo(String modulation, int distance) {
         this.modulation = modulation;
